Add entity framework core

+++ SimpleGantt.Domain/Queries/ProjectQueries.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 using System.Linq;
-using SimpleGantt.Domain.Entities.Common;
+using SimpleGantt.Domain.Entities;
 
 namespace SimpleGantt.Domain.Queries;
 
 public static class ProjectQueries
 {
-    public static long GetUsedResourcesAmountById(this Project project, Guid taskResourceId) =>
+    public static double GetUsedResourcesAmountById(this Project project, Guid taskResourceId) =>
         project.Tasks.Sum(item => item.Resources.First(taskResource => taskResource.Id == taskResourceId).Count);
 }
